requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.1.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coverage==7.10.6
cryptography==46.0.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
filelock==3.19.1
fsspec==2025.9.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-tools==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
huggingface-hub==0.35.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
openai==1.3.7
packaging==25.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
portalocker==2.10.1
protobuf==6.32.1
psutil==6.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyPDF2==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==7.0.0
python-docx==1.1.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pywin32==311
PyYAML==6.0.2
qdrant-client==1.7.0
regex==2025.9.1
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn==0.24.0
watchfiles==1.1.0
websockets==15.0.1
